refactor(tests): move screen tracking tests to dedicated test file
- Reverted MixpanelBasicTest.java to original state
- Created MixpanelScreenTrackingTest.java with all screenView/screenLeave tests
- Tests use InstrumentationRegistry for proper context access
